+++ lib/features/support/logger.rb
@@ -18,3 +18,44 @@ class MazeLogger < Logger
 end
 
 $logger = MazeLogger.instance
+
+# A collection of logging utilities
+class LogUtil
+  class << self
+    # Logs Hash-based data, accounting for things like file upload requests that are too big to log meaningfully.
+    #
+    # @param severity [Integer] A constant from Logger::Severity
+    # @param data [Hash] The data to log (currently needs to be a Hash)
+    def log_hash(severity, data)
+      return unless data.is_a? Hash
+
+      # Try to pretty print as JSON, if not too big
+      begin
+        json = JSON.pretty_generate data
+        if json.length < 128 * 1024
+          $logger.add severity, json
+        else
+          log_hash_by_field severity, data
+        end
+      rescue Encoding::UndefinedConversionError
+        log_hash_by_field severity, data
+      end
+    end
+
+    # Logs a hash field by field,
+    #
+    # @param severity [Integer] A Logger::Severity
+    # @param hash [Hash] The Hash
+    def log_hash_by_field(severity, hash)
+      hash.keys.each do |key|
+        value = hash[key].to_s
+        if value.length < 1024
+          $logger.add severity, "  #{key}: #{value}"
+        else
+          $logger.add severity, "  #{key} (length): #{value.length}"
+          $logger.add severity, "  #{key} (start): #{value[0, 1024]}"
+        end
+      end
+    end
+  end
+end